move(cosmo): move realizations to private API
