+++ astropy/cosmology/realizations.py
@@ -22,15 +22,11 @@ __all__ = [  # noqa: F822 (undefined name)
 
 import pathlib
 import sys
-from typing import TYPE_CHECKING
 
 from astropy.utils.data import get_pkg_data_path
-from astropy.utils.state import ScienceState
 
+from ._src.default import default_cosmology
 from .core import Cosmology
-
-if TYPE_CHECKING:
-    from typing import ClassVar
 
 __doctest_requires__ = {"*": ["scipy"]}
 
@@ -79,104 +75,3 @@ def __getattr__(name: str) -> Cosmology:
 def __dir__() -> list[str]:
     """Directory, including lazily-imported objects."""
     return __all__
-
-
-#########################################################################
-# The science state below contains the current cosmology.
-#########################################################################
-
-
-class default_cosmology(ScienceState):
-    """The default cosmology to use.
-
-    To change it::
-
-        >>> from astropy.cosmology import default_cosmology, WMAP7
-        >>> with default_cosmology.set(WMAP7):
-        ...     # WMAP7 cosmology in effect
-        ...     pass
-
-    Or, you may use a string::
-
-        >>> with default_cosmology.set('WMAP7'):
-        ...     # WMAP7 cosmology in effect
-        ...     pass
-
-    To get the default cosmology:
-
-        >>> default_cosmology.get()
-        FlatLambdaCDM(name='Planck18', H0=<Quantity 67.66 km / (Mpc s)>,
-                      Om0=0.30966, ...
-    """
-
-    _default_value: ClassVar[str] = "Planck18"
-    _value: ClassVar[str | Cosmology] = "Planck18"
-
-    @classmethod
-    def validate(cls, value: Cosmology | str | None) -> Cosmology | None:
-        """Return a Cosmology given a value.
-
-        Parameters
-        ----------
-        value : None, str, or `~astropy.cosmology.Cosmology`
-
-        Returns
-        -------
-        `~astropy.cosmology.Cosmology` instance
-
-        Raises
-        ------
-        TypeError
-            If ``value`` is not a string or |Cosmology|.
-        """
-        # None -> default
-        if value is None:
-            value = cls._default_value
-
-        # Parse to Cosmology. Error if cannot.
-        if isinstance(value, str):
-            # special-case one string
-            if value == "no_default":
-                value = None
-            else:
-                value = cls._get_from_registry(value)
-        elif not isinstance(value, Cosmology):
-            raise TypeError(
-                "default_cosmology must be a string or Cosmology instance, "
-                f"not {value}."
-            )
-
-        return value
-
-    @classmethod
-    def _get_from_registry(cls, name: str) -> Cosmology:
-        """Get a registered Cosmology realization.
-
-        Parameters
-        ----------
-        name : str
-            The built-in |Cosmology| realization to retrieve.
-
-        Returns
-        -------
-        `astropy.cosmology.Cosmology`
-            The cosmology realization of `name`.
-
-        Raises
-        ------
-        ValueError
-            If ``name`` is a str, but not for a built-in Cosmology.
-        TypeError
-            If ``name`` is for a non-Cosmology object.
-        """
-        try:
-            value = getattr(sys.modules[__name__], name)
-        except AttributeError:
-            raise ValueError(
-                f"Unknown cosmology {name!r}. Valid cosmologies:\n{available}"
-            )
-
-        if not isinstance(value, Cosmology):
-            raise TypeError(f"cannot find a Cosmology realization called {name}.")
-
-        return value
